refactor: externalize LoginForm component
